+++ bizapps-pesaje-web/src/interfacedao/maestros/ITransporteDAO.java
@@ -5,6 +5,6 @@ import modelo.maestros.Transporte;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ITransporteDAO extends  JpaRepository<Transporte, Integer>{
+public interface ITransporteDAO extends  JpaRepository<Transporte, Long>{
 
 }
